Optimize AES-GCM for uarchs with unroll and new instructions
Increase the block numbers to 8 for every iteration.
Increase the hash table capacity.
Make use of EOR3 instruction to improve the performance.

This can improve performance 25-40% on out-of-order
microarchitectures with a large number of fast execution units,
such as Neoverse V1.
We also see 20-30% performance improvements on other architectures
such as the M1.

Change-Id: I275d073e537fe6d4c396521be68ff4cf3ec86f56
CustomizedGitHooks: yes